+++ modules/app_lb/main.tf
@@ -1,7 +1,3 @@
-resource "google_compute_global_address" "default" {
-  name = "${var.namespace}-address"
-}
-
 resource "google_compute_global_address" "operator" {
   name = "${var.namespace}-operator-address"
 }
@@ -9,10 +5,9 @@ resource "google_compute_global_address" "operator" {
 module "https" {
   count = var.ssl ? 1 : 0
 
-  source     = "./https"
-  fqdn       = var.fqdn
-  namespace  = var.namespace
-  ip_address = google_compute_global_address.default.address
+  source    = "./https"
+  fqdn      = var.fqdn
+  namespace = var.namespace
 
   labels = var.labels
 }