+++ modules/kms/variables.tf
@@ -21,6 +21,12 @@ variable "bind_pubsub_service_to_kms_key" {
   default     = true
 }
 
+variable "bind_bigtable_service_to_kms_key" {
+  type        = bool
+  description = "Whether to bind the Pub/Sub service account to the KMS key for encrypter/decrypter access."
+  default     = true
+}
+
 variable "labels" {
   description = "Labels which will be applied to all applicable resources."
   type        = map(string)